requirements.txt
# Core dependencies
numpy>=1.26.4
pandas>=2.2.1
plotly>=5.19.0
plotly-express>=0.4.1

# Web framework
streamlit>=1.32.0

# Data processing
scikit-learn>=1.4.1
ta>=0.11.0
yfinance>=0.2.36

# Visualization
matplotlib>=3.8.3
seaborn>=0.13.2

# Utilities
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Database
pymongo>=4.6.1
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9

# Document processing
fpdf2>=2.7.7
pillow>=10.2.0
PyPDF2>=3.0.1
python-docx>=1.1.0 